python_service/requirements.txt
# Python Requirements for PDF Viewer Application

# Core PDF processing
PyMuPDF==1.24.1
Pillow==10.2.0

# Web framework for Python API
Flask==3.0.2
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# File handling and utilities
python-multipart==0.0.9
Werkzeug==3.0.1

# Development and testing
pytest==8.0.1
pytest-flask==1.3.0